python3-requests: Support idna version 3.1
Change idna required version to use python3-idna recipe that currently
is on version 3.1.

Without this change python-request doesn't work as the idna version
doesn't match the required versions. When bump python-request recipe,
verify if the upstream project bumped the idna version and drop this
patch.

As this recipe was tested with python3-idna_3.1 don't allow higher
versions.

Signed-off-by: Fabio Berton <fabio.berton@ossystems.com.br>
Signed-off-by: Khem Raj <raj.khem@gmail.com>