udpcast: add recipe
udpcast is a file transfer tool that can send data simultaneously to many
destinations on a LAN.

It is an alternative to uftp which is already in meta-networking.

On a lossy network, udpcast provided 10x faster transfer rates
with error correction enabled while using a half as much CPU thanks to a
simpler algorithm.

Signed-off-by: Khem Raj <raj.khem@gmail.com>